Directly use `Node` to avoid `List` object allocation
